Define embulkSetup as inline
